pyproject.toml
[project]
name = "resume-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "boto3>=1.40.18",
    "fastapi[standard]>=0.116.1",
    "python-decouple>=3.8",
    "sqlalchemy[asyncio]>=2.0.43",
]
